Move browser sources to src subfolder
