+++ browser/src/error.ts
@@ -1,4 +1,4 @@
-import { errNoFileSystemInBrowser, error } from '../src/utils/error';
+import { errNoFileSystemInBrowser, error } from '../../src/utils/error';
 
 export const throwNoFileSystem = (method: string) => (): never =>
 	error(errNoFileSystemInBrowser(method));
