+++ browser/src/package.json
@@ -2,8 +2,8 @@
   "name": "@rollup/browser",
   "version": "3.0.0-2",
   "description": "Next-generation ES module bundler browser build",
-  "main": "dist/rollup.browser.js",
-  "module": "dist/es/rollup.browser.js",
+  "main": "../dist/rollup.browser.js",
+  "module": "../dist/es/rollup.browser.js",
   "types": "dist/rollup.browser.d.ts",
   "repository": "rollup/rollup",
   "keywords": [
@@ -22,7 +22,7 @@
   "files": [
     "dist/**/*.js",
     "dist/*.d.ts",
-    "dist/es/package.json",
+    "../dist/es/package.json",
     "dist/rollup.browser.js.map"
   ],
   "exports": {
