+++ browser/src/path.ts
