+++ browser/src/performance.ts
